Finally got it working
