+++ IronScheme/IronScheme.Closures/IronScheme.Closures.csproj
@@ -32,8 +32,8 @@
     <ProjectReference Include="..\Microsoft.Scripting\IronScheme.Scripting.csproj" />
   </ItemGroup>
   <ItemGroup Condition="$(UseNetFxIlTools) != True">
-    <PackageReference Include="runtime.win-x64.Microsoft.NETCore.ILAsm" Version="9.0.0-preview.7.24405.7" GeneratePathProperty="true" ExcludeAssets="all" />
-    <PackageReference Include="runtime.win-x64.Microsoft.NETCore.ILDAsm" Version="9.0.0-preview.7.24405.7" GeneratePathProperty="true" ExcludeAssets="all" />
+    <PackageReference Include="runtime.win-x64.Microsoft.NETCore.ILAsm" Version="9.0.0-*" GeneratePathProperty="true" ExcludeAssets="all" />
+    <PackageReference Include="runtime.win-x64.Microsoft.NETCore.ILDAsm" Version="9.0.0-*" GeneratePathProperty="true" ExcludeAssets="all" />
   </ItemGroup>
   <ItemGroup>
     <Content Include="IronScheme.Glue.il">
