+++ IronScheme/Microsoft.Scripting/IPAL.cs
@@ -6,6 +6,14 @@ using System.Reflection.Emit;
 
 namespace IronScheme.FrameworkPAL
 {
+    public interface ISerializer
+    {
+        object Deserialize(Stream serializationStream);
+        void Serialize(Stream serializationStream, object graph);
+    }
+
+    public delegate Type RecordBinderCallback(string assName, string typeName);
+
     public interface IPAL
     {
         void Initialize();
@@ -18,11 +26,13 @@ namespace IronScheme.FrameworkPAL
 
         void DefineAssembly(bool run, string outDir, AssemblyName asmname, string actualModuleName, string outFileName, bool emitDebugInfo, ref AssemblyBuilder ab, ref ModuleBuilder mb);
         void SerializeConstants(MemoryStream s, ModuleBuilder mb, bool compress);
+        ISerializer GetSerializer(RecordBinderCallback rb);
     }
 
     public static class PAL
     {
         static readonly bool IsCore = typeof(object).Assembly.FullName.StartsWith("System.Private.CoreLib");
+        static readonly bool IsNet9 = IsCore && typeof(object).Assembly.GetName().Version.Major >= 9;
 
         static readonly IPAL pal = LoadPAL();
 
@@ -40,7 +50,7 @@ namespace IronScheme.FrameworkPAL
             }
             else
             {
-                var resource = typeof(PAL).Assembly.GetManifestResourceStream($"{(IsCore ? "core-" : "")}{fn}");
+                var resource = typeof(PAL).Assembly.GetManifestResourceStream($"{(IsCore ? (IsNet9 ? "net9-": "core-") : "")}{fn}");
                 if (resource == null)
                 {
                     throw new PlatformNotSupportedException();
@@ -62,8 +72,9 @@ namespace IronScheme.FrameworkPAL
         public static void SerializeConstants(MemoryStream s, ModuleBuilder mb, bool compress) => pal.SerializeConstants(s, mb, compress);
         public static void Save(AssemblyBuilder ass, string filename, ImageFileMachine machineKind) => pal.Save(ass, filename, machineKind);
         public static void SetLocalSymInfo(LocalBuilder lb, string name) => pal.SetLocalSymInfo(lb, name);
-
         public static void DefineAssembly(bool run, string outDir, AssemblyName asmname, string actualModuleName, string outFileName, bool emitDebugInfo, ref AssemblyBuilder ab, ref ModuleBuilder mb) =>
             pal.DefineAssembly(run, outDir, asmname, actualModuleName, outFileName, emitDebugInfo, ref ab, ref mb);
+
+        public static ISerializer GetSerializer(RecordBinderCallback recordBinder) => pal.GetSerializer(recordBinder);
     }
 }
